changed

+++ src/Flowers/parent/Item.java
@@ -3,5 +3,5 @@ package Flowers.parent;
 
 public interface Item {
     public String getDescription();
-    public double price();
+    public int price();
 }
